pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hologram-lab"
version = "0.1.0"
description = "A comprehensive repository for hologram creation, simulations, and resources"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "ElaMereanu", email = ""}
]
keywords = ["hologram", "computer-generated-holography", "cgh", "slm", "optics", "holography"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.4.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "pillow>=8.3.0",
    "imageio>=2.9.0",
    "scikit-image>=0.18.0",
    "opencv-python>=4.5.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
holo-sim = "hologram_lab.cli:main"

[tool.setuptools]
packages = ["hologram_lab"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["simulations/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

